Edited Generator to take in Doubles, and edited all of the HashMaps to be <_, Double>.

+++ Generator.java
@@ -6,7 +6,7 @@ public class Generator {
 	/*
 	 * Generates a random sentence based on a unigram language model.
 	 */
-	public static String randomUnigramSentence(HashMap<Token,Integer> unigrams) {
+	public static String randomUnigramSentence(HashMap<Token,Double> unigrams) {
         //int total = getTotal(unigrams);
         String sentence = "";
         boolean isEnd = false; //boolean indicating whether sentence is done
@@ -35,13 +35,13 @@ public class Generator {
 	 * Selects a Token with probability according to its corresponding value 
 	 * with roulette style selection
 	 */
-    public static Token selectUnigramWithProbability(HashMap<Token,Integer> model) {
+    public static Token selectUnigramWithProbability(HashMap<Token,Double> model) {
 		int sum = 0;
-		for(Integer count: model.values()) {
+		for(Double count: model.values()) {
 			sum += count;
 		}
 		
-		int rand = (int) (Math.random() * sum);
+		double rand = (Math.random() * sum);
 		int cur = 0;
 		for(Token t: model.keySet()) {
 			cur += model.get(t);
@@ -56,14 +56,14 @@ public class Generator {
 	/*
 	 * Generates a random sentence based on a bigram language model.
 	 */
-	public static String randomBigramSentence(HashMap<Bigram, Integer> bigram_model) {
+	public static String randomBigramSentence(HashMap<Bigram, Double> bigram_model) {
 		
 		ArrayList<Bigram> sentence_arr = new ArrayList<>();
 		Token prev_word = null;
 		
 		while(prev_word == null || prev_word.getType() != TokenType.END) {
 			// Construct map of bigrams matching the previous word
-			HashMap<Bigram, Integer> possible_bigrams = new HashMap<>();
+			HashMap<Bigram, Double> possible_bigrams = new HashMap<>();
 			for(Bigram b: bigram_model.keySet()) {
 				if(prev_word == null) {
 					if(b.getFirst().getType() == TokenType.START) {
@@ -103,13 +103,13 @@ public class Generator {
 	 * The number of occurrences of the preceding word should be the same as the 
 	 * sum of all values in the model received
 	 */
-	private static Bigram selectBigramWithProbability(HashMap<Bigram, Integer> model) {
+	private static Bigram selectBigramWithProbability(HashMap<Bigram, Double> model) {
 		int sum = 0;
-		for(Integer count: model.values()) {
+		for(Double count: model.values()) {
 			sum += count;
 		}
 		
-		int rand = (int) (Math.random() * sum);
+		double rand = (Math.random() * sum);
 		int cur = 0;
 		for(Bigram b: model.keySet()) {
 			cur += model.get(b);
@@ -125,50 +125,50 @@ public class Generator {
 		// "The grey fox likes cats.", "The grey fox hates dogs.", "The red fox likes Pokemon."
 		// "The grey deer likes cookies.", "The grey fox likes DeadMau5.", "The red deer hates Muppets."
 		
-		HashMap<Token, Integer> unigram_model = new HashMap<>();
+		HashMap<Token, Double> unigram_model = new HashMap<>();
 		
-		unigram_model.put(new Token("The", TokenType.WORD), 6);
-		unigram_model.put(new Token("grey", TokenType.WORD), 4);
-		unigram_model.put(new Token("red", TokenType.WORD), 2);
-		unigram_model.put(new Token("fox", TokenType.WORD), 4);
-		unigram_model.put(new Token("deer", TokenType.WORD), 2);
-		unigram_model.put(new Token("likes", TokenType.WORD), 4);
-		unigram_model.put(new Token("hates", TokenType.WORD), 2);
-		unigram_model.put(new Token("cats", TokenType.WORD), 1);
-		unigram_model.put(new Token("dogs", TokenType.WORD), 1);
-		unigram_model.put(new Token("Pokemon", TokenType.WORD), 1);
-		unigram_model.put(new Token("cookies", TokenType.WORD), 1);
-		unigram_model.put(new Token("DeadMau5", TokenType.WORD), 1);
-		unigram_model.put(new Token("Muppets", TokenType.WORD), 1);
-		unigram_model.put(new Token(null, TokenType.START), 6);
-		unigram_model.put(new Token(".", TokenType.WORD), 6);
-		unigram_model.put(new Token(null, TokenType.END), 6);
+		unigram_model.put(new Token("The", TokenType.WORD), 6.0);
+		unigram_model.put(new Token("grey", TokenType.WORD), 4.0);
+		unigram_model.put(new Token("red", TokenType.WORD), 2.0);
+		unigram_model.put(new Token("fox", TokenType.WORD), 4.0);
+		unigram_model.put(new Token("deer", TokenType.WORD), 2.0);
+		unigram_model.put(new Token("likes", TokenType.WORD), 4.0);
+		unigram_model.put(new Token("hates", TokenType.WORD), 2.0);
+		unigram_model.put(new Token("cats", TokenType.WORD), 1.0);
+		unigram_model.put(new Token("dogs", TokenType.WORD), 1.0);
+		unigram_model.put(new Token("Pokemon", TokenType.WORD), 1.0);
+		unigram_model.put(new Token("cookies", TokenType.WORD), 1.0);
+		unigram_model.put(new Token("DeadMau5", TokenType.WORD), 1.0);
+		unigram_model.put(new Token("Muppets", TokenType.WORD), 1.0);
+		unigram_model.put(new Token(null, TokenType.START), 6.0);
+		unigram_model.put(new Token(".", TokenType.WORD), 6.0);
+		unigram_model.put(new Token(null, TokenType.END), 6.0);
 
-		HashMap<Bigram, Integer> bigram_model = new HashMap<>();
-		bigram_model.put(new Bigram(new Token(null, TokenType.START), new Token("The", TokenType.WORD)), 4);
-		bigram_model.put(new Bigram(new Token("The", TokenType.WORD), new Token("grey", TokenType.WORD)), 4);
-		bigram_model.put(new Bigram(new Token("The", TokenType.WORD), new Token("red", TokenType.WORD)), 2);
-		bigram_model.put(new Bigram(new Token("grey", TokenType.WORD), new Token("fox", TokenType.WORD)), 3);
-		bigram_model.put(new Bigram(new Token("grey", TokenType.WORD), new Token("deer", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("red", TokenType.WORD), new Token("fox", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("red", TokenType.WORD), new Token("deer", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("fox", TokenType.WORD), new Token("likes", TokenType.WORD)), 3);
-		bigram_model.put(new Bigram(new Token("fox", TokenType.WORD), new Token("hates", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("deer", TokenType.WORD), new Token("likes", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("deer", TokenType.WORD), new Token("hates", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("likes", TokenType.WORD), new Token("cats", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("likes", TokenType.WORD), new Token("Pokemon", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("likes", TokenType.WORD), new Token("cookies", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("likes", TokenType.WORD), new Token("DeadMau5", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("hates", TokenType.WORD), new Token("dogs", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("hates", TokenType.WORD), new Token("Muppets", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("cats", TokenType.WORD), new Token(".", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("Pokemon", TokenType.WORD), new Token(".", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("cookies", TokenType.WORD), new Token(".", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("DeadMau5", TokenType.WORD), new Token(".", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("dogs", TokenType.WORD), new Token(".", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token("Muppets", TokenType.WORD), new Token(".", TokenType.WORD)), 1);
-		bigram_model.put(new Bigram(new Token(".", TokenType.WORD), new Token(null, TokenType.END)), 6);
+		HashMap<Bigram, Double> bigram_model = new HashMap<>();
+		bigram_model.put(new Bigram(new Token(null, TokenType.START), new Token("The", TokenType.WORD)), 4.0);
+		bigram_model.put(new Bigram(new Token("The", TokenType.WORD), new Token("grey", TokenType.WORD)), 4.0);
+		bigram_model.put(new Bigram(new Token("The", TokenType.WORD), new Token("red", TokenType.WORD)), 2.0);
+		bigram_model.put(new Bigram(new Token("grey", TokenType.WORD), new Token("fox", TokenType.WORD)), 3.0);
+		bigram_model.put(new Bigram(new Token("grey", TokenType.WORD), new Token("deer", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("red", TokenType.WORD), new Token("fox", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("red", TokenType.WORD), new Token("deer", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("fox", TokenType.WORD), new Token("likes", TokenType.WORD)), 3.0);
+		bigram_model.put(new Bigram(new Token("fox", TokenType.WORD), new Token("hates", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("deer", TokenType.WORD), new Token("likes", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("deer", TokenType.WORD), new Token("hates", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("likes", TokenType.WORD), new Token("cats", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("likes", TokenType.WORD), new Token("Pokemon", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("likes", TokenType.WORD), new Token("cookies", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("likes", TokenType.WORD), new Token("DeadMau5", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("hates", TokenType.WORD), new Token("dogs", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("hates", TokenType.WORD), new Token("Muppets", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("cats", TokenType.WORD), new Token(".", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("Pokemon", TokenType.WORD), new Token(".", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("cookies", TokenType.WORD), new Token(".", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("DeadMau5", TokenType.WORD), new Token(".", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("dogs", TokenType.WORD), new Token(".", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token("Muppets", TokenType.WORD), new Token(".", TokenType.WORD)), 1.0);
+		bigram_model.put(new Bigram(new Token(".", TokenType.WORD), new Token(null, TokenType.END)), 6.0);
 
 		String sentence = Generator.randomBigramSentence(bigram_model);
 		System.out.println(sentence);
